Add switch to control simulator / REST context
