+++ src/server/context/rest.js
@@ -1,0 +1,4 @@
+// Use the OpenStack client to implement the context API.
+const context = {}
+
+export default context
